reorg with cli.utils module
